fix interceptor to refresh token

+++ frontend/src/app/views/auth/services/auth-interceptor.service.ts
@@ -5,11 +5,12 @@ import {
   HttpRequest,
 } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Observable, catchError, concatMap, from, of, throwError } from 'rxjs';
+import { Observable, concatMap, of } from 'rxjs';
 import { JwtTokenService } from './jwt-token.service';
 import { SnackbarService } from '../../../shared/services/snackbar.service';
 import { Router } from '@angular/router';
 import { TokenResponse } from '../models/auth';
+import { environment } from '../../../../environments/environment';
 
 @Injectable({
   providedIn: 'root',
@@ -25,13 +26,19 @@ export class AuthInterceptorService implements HttpInterceptor {
     req: HttpRequest<any>,
     next: HttpHandler
   ): Observable<HttpEvent<any>> {
+    if (req.url === environment.authentication.refreshToken) {
+      return next.handle(req);
+    }
+
     if (this.jwtTokenService.isAccessTokenExpired()) {
       if (this.jwtTokenService.isRefreshTokenExpired()) {
         this.jwtTokenService.removeToken();
         this.router.navigate(['/auth/sign-in']);
         this.snackBarService.openSnackBar('Sesja wygasła', true);
+
+        return of();
       } else {
-        this.jwtTokenService.refreshTokens().pipe(
+        return this.jwtTokenService.refreshTokens().pipe(
           concatMap((token) => {
             const newReq = this.addAuthHeader(req, token);
 
@@ -39,19 +46,19 @@ export class AuthInterceptorService implements HttpInterceptor {
           })
         );
       }
-    } else {
-      const token = this.jwtTokenService.getToken();
-      const newReq = token ? this.addAuthHeader(req, token) : req;
-
-      return next.handle(newReq);
     }
+    
+    const token = this.jwtTokenService.getToken();
+    const newReq = token ? this.addAuthHeader(req, token) : req;
 
-    return next.handle(req);
+    return next.handle(newReq);
   }
 
   addAuthHeader(req: HttpRequest<any>, token: TokenResponse) {
     return token
-      ? req.clone({ setHeaders: { Authorization: `Bearer ${token.accessToken}` } })
+      ? req.clone({
+          setHeaders: { Authorization: `Bearer ${token.accessToken}` },
+        })
       : req;
   }
 }
